Register jinja short_datetime filter for all apps
