+++ sonic_app.py
@@ -86,27 +86,11 @@ app.config["SOLFLARE_PATH"] = os.getenv("SOLFLARE_PATH", "C:/alpha/wallets/solfl
 socketio = SocketIO(app)
 
 # --- Template Filters ---
+from utils.template_filters import short_datetime
 
-@app.template_filter("short_datetime")
-def short_datetime(value):
-    """Format timestamps like '2025-06-03T13:06:03.968905' as '3:06PM 6/3/25'."""
-    if not value:
-        return ""
-    try:
-        if isinstance(value, (int, float)):
-            dt = datetime.fromtimestamp(float(value))
-        else:
-            try:
-                dt = datetime.fromisoformat(str(value))
-            except ValueError:
-                dt = datetime.strptime(str(value), "%Y-%m-%d %H:%M:%S")
-        formatted = dt.strftime("%I:%M%p %m/%d/%y")
-        if formatted.startswith("0"):
-            formatted = formatted[1:]
-        formatted = formatted.replace("/0", "/")
-        return formatted
-    except Exception:
-        return value
+# Register filters with the Flask app so they are available even when this file
+# is imported in isolation (e.g. during tests).
+app.add_template_filter(short_datetime, "short_datetime")
 
 # --- SINGLETON BACKEND ---
 app.data_locker = DataLocker(str(DB_PATH))
